requirements.txt
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==2.15.8
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.3.0
bleach==6.1.0
blis==0.7.11
boto3==1.35.20
botocore==1.35.20
catalogue==2.0.10
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
confection==0.1.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==43.0.1
cymem==2.0.8
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2.16
django-admin-sortable2==2.2.3
django-cache-memoize==0.2.0
django-cors-headers==4.4.0
django-coverage-plugin==2.0.4
django-debug-toolbar==4.4.6
django-dynamic-preferences==1.16.0
django-easy-health-check==1.0.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.1.1
django-prettyjson==0.4.1
django-stubs==1.12.0
django-stubs-ext==5.0.4
django-taggit==6.0.0
django-timezone-field==7.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-stubs==1.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.2.2
executing==2.1.0
factory_boy==3.3.1
Faker==28.4.1
fastjsonschema==2.20.0
filelock==3.16.0
flake8==7.1.1
flake8-isort==6.1.1
fqdn==1.5.1
gevent==24.2.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.9
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kombu==5.4.1
langcodes==3.4.0
language_data==1.2.0
lazy-object-proxy==1.10.0
marisa-trie==1.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
murmurhash==1.0.10
mypy==0.991
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.25.0
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathlib_abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
persisting-theory==1.0
pexpect==4.9.0
phonenumbers==8.13.45
platformdirs==4.3.3
pluggy==1.5.0
pre-commit==2.21.0
preshed==3.0.9
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytest==7.3.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-mock==1.12.1
respx==0.21.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
s3transfer==0.10.2
Send2Trash==1.8.3
sentry-sdk==2.14.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
standardjson==0.3.1
tablib==3.5.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.1.12
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.9.4
types-python-dateutil==2.9.0.20240906
types-pytz==2024.2.0.20240913
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
virtualenv==20.26.4
wasabi==1.1.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.7.0
widgetsnbextension==4.0.13
wrapt==1.16.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.0.3
